refactor: extract checkbox props to separate file
