+++ src/Shared/Components/CheckBox/index.ts
@@ -1,1 +1,2 @@
 export * from "./CheckBox";
+export * from "./propsCheckBox";
