+++ src/Shared/Components/CheckBox/propsCheckBox.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+export interface CheckBoxProps
+  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, "onChange"> {
+  checked?: boolean | undefined;
+  disabled?: boolean | undefined;
+  onChange: (value: boolean) => void;
+  className?: string;
+}
